最终版

+++ Jenkinsfile
@@ -3,13 +3,14 @@ import groovy.json.JsonSlurper
 node {
 
     // 需要手动赋值的变量(是否需要全部移到config.json)
-    def appointWorkDir = false                        // 是否指定挂载目录
-    def projectName = 'jenkins-test'                  // 项目名称
-    def hostWorkDir = '/var/jenkins/jenkins-test'     // 宿主机目录(随便设置)
-    def workDir = '/var/jenkins/jenkins-test'         // 容器目录(Dockerfile中的WORKDIR目录)
-    def openPort = '7777'                             // 开发的端口(Dockerfile中的EXPOSE,使用该端口映射)
-    def otherArgs = "--name ${projectName}"           // 自定义其他启动参数(默认启动参数格式为: docker run -d -p 8080:8080 -e pro_env=dev)
-    def mvnCMD = "mvn -B -DskipTests clean package"   // maven打包命令 暂时先不用这个字段
+    def appointWorkDir = false                              // 是否指定挂载目录
+    def projectName = 'jenkins-test'                        // 项目名称
+    def hostWorkDir = '/var/jenkins/jenkins-test'           // 宿主机目录(随便设置)
+    def workDir = '/var/jenkins/jenkins-test'               // 容器目录(Dockerfile中的WORKDIR目录)
+    def openPort = '7777'                                   // 开放的端口(Dockerfile中的EXPOSE,使用该端口映射)
+    def otherArgs = "--name ${projectName}"                 // 自定义其他启动参数(默认启动参数格式为: docker run -d -p 8080:8080 -e pro_env=dev)
+    def buildArgs = " --build-arg PRO_ENV=${env.PRO_ENV} ." // 自定义其他构建镜像参数(默认参数格式为: docker build -t imageName:tag .)
+    def mvnCMD = "mvn -B -DskipTests clean package"         // maven打包命令
 
 
     // 系统变量
@@ -25,7 +26,7 @@ node {
     currentBuild.result = "SUCCESS"
 
     try {
-        stage('config') {
+        stage('choice') {
 
             inputResult = input message: '请选择部署环境', ok: '确认', parameters: [
                 booleanParam(name: 'dev', defaultValue: true),
@@ -39,7 +40,7 @@ node {
             if (inputResult.test) {
                 env.PRO_ENV = "test"
             }
-            if (inputResult.deploy) {
+            if (inputResult.prod) {
                 env.PRO_ENV = "prod"
             }
             sh 'env $env'
@@ -51,9 +52,9 @@ node {
             // checkout([$class: 'GitSCM', branches: [[name: '*/dev']], extensions: [], userRemoteConfigs: [[credentialsId: 'e761309f-146f-4c5f-b7fd-1debf91fef38', url: 'https://gitee.com/original-blackhole/jenkins-test.git']]])
             withEnv(["MVN_HOME=$mvnHome"]) {
                 if (isUnix()) {
-                    sh 'mvn -B -DskipTests clean package'
+                    sh "${mvnCMD}"
                 } else {
-                    bat(/mvn -B -DskipTests clean package/)
+                    bat(/${mvnCMD}/)
                 }
             }
 
@@ -64,13 +65,13 @@ node {
                 def obj = jsonSlurper.parseText(str)
                 echo "config.json: ${str}"
 
+                env.VERSION = obj.version
                 env.registryName = obj.registryName
+                env.registry_credentials_id = obj.registry_credentials_id
                 def envConifg = obj.env[env.PRO_ENV]
                 echo "envConifg: ${envConifg}"
 
-                env.VERSION = obj.version
                 env.credentialsId = envConifg.credentialsId
-                env.registry_credentials_id = envConifg.registry_credentials_id
                 env.host = envConifg.host
                 env.port = envConifg.port
                 env.name = envConifg.name
@@ -84,8 +85,9 @@ node {
 
             // 构建镜像并上传到容器仓库
             echo "imageName: ${imageName}"
-            def customImage = docker.build(imageName, " --build-arg pro_env=${env.PRO_ENV} .")
+            def customImage = docker.build(imageName, " --build-arg PRO_ENV=${env.PRO_ENV} .")
             if(!(inputResult.dev)) {
+                echo "registry_credentials_id: ${env.registry_credentials_id} "
                 docker.withRegistry("https://${env.registryName}", env.registry_credentials_id) {
                     customImage.push()
                 }
@@ -93,18 +95,20 @@ node {
             sh 'env $env'
         }
 
-        stage('Deploy'){
+        stage('deploy'){
             if(inputResult.dev) {
 
                 // 本地部署
                 echo "start local deploy, host: ${env.host}"
                 try {
-                    // 停止删除容器
-                    // sh "docker rm \$(docker stop \$(docker ps -a | grep ${projectName} | awk '{print \$1}'))"  //没有容器时会报错
-                    // 停止容器
-                    sh "docker stop ${projectName}"   //TODO 没有容器时会报错 但是不影响执行流程先不改了
-                    // 删除容器
-                    sh "docker rm ${projectName}"
+
+                    result = sh(script: "docker ps -a | grep ${projectName} | awk '{print \$1}'", returnStdout: true).trim()
+                    echo "result : ${result}"
+                    if(result){
+                        sh "docker stop ${projectName}"
+                        sh "docker rm -f ${projectName}"
+                    }
+
                 } catch (err) {}
                 // 启动容器
                 def serverArgs = "-p ${env.port}:${openPort} --name ${projectName}"
@@ -119,36 +123,37 @@ node {
 
                     // 部署集群
                     //for (item in ipList.tokenize(',')){
-                        def sshServer = getServer(env.credentialsId,env.host,false)
+                        def sshServer = getServer(env.credentialsId,env.host,false,projectName)
 
                         // 更新或下载镜像
                         sshCommand remote: sshServer, command: "docker pull ${imageName}"
 
                         try{
                             // 停止容器
-                            sshCommand remote: sshServer, command: "docker stop ${projectName}"
-                            // 删除容器
-                            sshCommand remote: sshServer, command: "docker rm -f ${projectName}"
-                            // sshCommand remote: remote, command: "docker rm \$(docker stop \$(docker ps -a | grep ${projectName} | awk '{print \$1}'))"
+                            def stopSSH = sshCommand remote: sshServer, command: "docker ps -a | grep ${projectName} | awk '{print \$1}'"
+                            if(stopSSH){
+                                sshCommand remote: sshServer, command: "docker stop ${projectName}"
+                                sshCommand remote: sshServer, command: "docker rm -f ${projectName}"
+                            }
 
-                        }catch(ex){}
+                        }catch(err){}
 
                         // 启动参数
-                        def serverArgs = "docker run -d -p ${env.port}:${openPort} -e pro_env=${env.PRO_ENV}"
+                        def serverArgs = "docker run -d -p ${env.port}:${openPort} -e PRO_ENV=${env.PRO_ENV}"
                         if(appointWorkDir){
                             serverArgs = "${serverArgs} -v ${hostWorkDir}:${openPort}"
                         }
-                        echo 'serverArgs: ${serverArgs} ${otherArgs} ${imageName}'
+                        echo "serverArgs: ${serverArgs} ${otherArgs} ${imageName}"
 
                         // 启动容器
-                        sshCommand remote: sshServer, command: "${serverArgs} ${otherArgs} $imageName}"
+                        sshCommand remote: sshServer, command: "${serverArgs} ${otherArgs} ${imageName}"
 
                         // 删除远程服务器的上两个版本镜像(只保留最新的两个版本镜像)
                         sshCommand remote: sshServer, command: "docker rmi -f ${imageName.replaceAll("_${BUILD_NUMBER}", "_${BUILD_NUMBER.toInteger() - 1}")}"
                         // 清理none镜像
-                        // def clearNoneSSH = "n=`docker images | grep  '<none>' | wc -l`; if [ \$n -gt 0 ]; then docker rmi `docker images | grep  '<none>' | awk '{print \$3}'`; fi"
-                        // sshCommand remote: sshServer, command: "${clearNoneSSH}"
-                        sshCommand remote: sshServer, command: "docker rmi -f \$(docker images | grep none | awk '{print \$3}')"
+                        def clearNoneSSH = "n=`docker images | grep  '<none>' | wc -l`; if [ \$n -gt 0 ]; then docker rmi `docker images | grep  '<none>' | awk '{print \$3}'`; fi"
+                        sshCommand remote: sshServer, command: "${clearNoneSSH}"
+                        // sshCommand remote: sshServer, command: "docker rmi -f \$(docker images | grep none | awk '{print \$3}')"
                     //}
 
 
@@ -168,23 +173,10 @@ node {
         throw err
     }
 }
-// 声明一个获取服务的函数,我这里所有服务的密码都是一样的,如有区别可自行修改函数
-def getServer(ip){
+// 声明一个获取服务的函数(默认端口22)
+def getServer(credentialsId,ip,port,projectName){
     def remote = [:]
-    remote.name = "server-${ip}"
-    remote.host = ip
-    remote.port = 22
-    remote.allowAnyHosts = true
-    withCredentials([usernamePassword(credentialsId: '8a0eadde-27ec-424f-97e3-aefcf54e5caf', passwordVariable: 'password', usernameVariable: 'userName')]) {
-    //withCredentials([sshUserPrivateKey(credentialsId: 'ServiceServer', keyFileVariable: 'identity', passphraseVariable: '', usernameVariable: 'userName')]) {
-        remote.user = "${userName}"
-        remote.password = "${password}"
-    }
-    return remote
-}
-def getServer(credentialsId,ip,port){
-    def remote = [:]
-    remote.name = "server-${ip}"
+    remote.name = "${projectName}"
     remote.host = ip
     remote.port = 22
     remote.allowAnyHosts = true
